Refactor VersionPage tests into separate test classes
